+++ src/events/index.ts
@@ -4,3 +4,4 @@ export * from "./detectors";
 export * from "./cooldowns";
 export * from "./select";
 export * from "./resolve";
+export * from "./resolve-social";